requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via quart
argon2-cffi==23.1.0
    # via -r requirements.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
black==23.11.0
    # via -r requirements.in
blinker==1.7.0
    # via
    #   flask
    #   quart
cffi==1.16.0
    # via argon2-cffi-bindings
click==8.1.7
    # via
    #   black
    #   flask
    #   quart
flask==3.0.0
    # via quart
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.15.0
    # via quart
hyperframe==6.0.1
    # via h2
isort==5.12.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via
    #   flask
    #   quart
jinja2==3.1.2
    # via
    #   flask
    #   quart
markupsafe==2.1.3
    # via
    #   -r requirements.in
    #   jinja2
    #   quart
    #   werkzeug
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via black
pathspec==0.11.2
    # via black
platformdirs==4.1.0
    # via black
priority==2.0.0
    # via hypercorn
pycparser==2.21
    # via cffi
quart==0.19.4
    # via -r requirements.in
sqlalchemy==2.0.23
    # via -r requirements.in
typing-extensions==4.8.0
    # via sqlalchemy
werkzeug==3.0.1
    # via
    #   flask
    #   quart
wsproto==1.2.0
    # via hypercorn
